--- src/main/java/com/specialwarriors/conal/common/jwt/JwtFilter.java
@@ -1,63 +1,0 @@
-package com.specialwarriors.conal.common.jwt;
-
-import com.specialwarriors.conal.common.auth.CustomUserDetailsService;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class JwtFilter extends OncePerRequestFilter {
-
-    private final JwtProvider jwtProvider;
-    private final CustomUserDetailsService userDetailsService;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain filterChain)
-            throws ServletException, IOException {
-
-        if (request.getRequestURI().equals("/")) { // TODO : 기능 추가되면 인증 필요없는 URI 추가
-            filterChain.doFilter(request, response);
-        }
-
-        String token = resolveToken(request);
-
-        if (jwtProvider.validateToken(token)) {
-            Long userId = jwtProvider.getUserId(token);
-
-            UserDetails userDetails = userDetailsService.loadUserByUserId(userId);
-
-            UsernamePasswordAuthenticationToken authentication =
-                    new UsernamePasswordAuthenticationToken(userDetails, null,
-                            userDetails.getAuthorities());
-
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String token = request.getHeader("Authorization");
-
-        if (token == null || token.isEmpty()) {
-            throw new RuntimeException(); // TODO : 예외 처리
-        }
-        if (token.startsWith("Bearer ")) {
-            return token.substring(7);
-        }
-        return null;
-    }
-}
